requirements.txt
# local package
-e .

# external requirements
pandas
numpy
matplotlib
scikit-learn
awscli
xgboost
mlflow<3
dagshub
ipykernel
nltk
PyYAML